Rename and move into subdirectory js_only
