+++ assets/prototype/application/services/apollo/relations/test/useRelationsManager/getData.test.ts
@@ -1,7 +1,7 @@
 import { renderHook, act } from '@testing-library/react-hooks';
 
-import { relationalData } from '../../../../../domain/eventEditor/context';
-import useRelationsManager from '../useRelationsManager';
+import { relationalData } from '../../../../../../domain/eventEditor/context';
+import useRelationsManager from '../../useRelationsManager';
 
 describe('RelationsManager.getData()', () => {
 	it('returns the initial relational state (empty object)', () => {
