+++ assets/prototype/application/services/apollo/relations/test/useRelationsManager/setData.test.ts
@@ -1,8 +1,8 @@
 import { renderHook, act } from '@testing-library/react-hooks';
 import { pick } from 'ramda';
 
-import { relationalData } from '../../../../../domain/eventEditor/context';
-import useRelationsManager from '../useRelationsManager';
+import { relationalData } from '../../../../../../domain/eventEditor/context';
+import useRelationsManager from '../../useRelationsManager';
 
 describe('RelationsManager.setData()', () => {
 	it('returns the relational state set via setData when initial state is NOT passed', () => {
